apm0013173-css-ra-utilities/requirements.txt
pip

Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
oracledb==3.4.1
gunicorn

setuptools==78.1.1
requests
coverage
xmltodict
comment-parser
ntc-templates
netmiko
xlsxwriter
azure-identity
azure-storage-blob
azure-keyvault-secrets